Доработки
Доработки

+++ CarTek.Api/Model/Dto/FillDocumentModel.cs
@@ -45,10 +45,10 @@ namespace CarTek.Api.Model.Dto
         public Unit? UnloadUnit2 { get; set; }
 
         //Id точки забора груза
-        public int? LocationAId { get; set; }
+        public long? LocationAId { get; set; }
 
         //Id точки сдачи груза
-        public int? LocationBId { get; set; }
+        public long? LocationBId { get; set; }
 
         public DateTime? PickUpDepartureDate { get; set; }
